fix: addSeriesArticles 返回类型改 Long 避开 Result.ok(Integer errCode) 重载陷阱

+++ oy-blog-service/admin-service/src/main/java/com/oyproj/controller/AdminArticleController.java
@@ -98,7 +98,7 @@ public class AdminArticleController {
     @PostMapping("/series/{seriesId}/articles")
     @RequirePermission("admin:article:write")
     @Operation(summary = "批量收录文章进专栏")
-    public Result<Integer> addSeriesArticles(@PathVariable("seriesId") String seriesId,
+    public Result<Long> addSeriesArticles(@PathVariable("seriesId") String seriesId,
                                              @RequestBody SeriesMemberBindDto dto) {
         return biz.addSeriesArticles(seriesId, dto);
     }
